+++ src/App.cpp
@@ -33,7 +33,6 @@
 #include "common/Platform.h"
 #include "core/Config.h"
 #include "core/Controller.h"
-#include "Cpu.h"
 #include "crypto/CryptoNight.h"
 #include "net/Network.h"
 #include "Summary.h"
@@ -124,7 +123,7 @@ int App::exec()
 #   endif
 
     if (!m_controller->oclInit() || !Workers::start(m_controller)) {
-        LOG_ERR("Failed to start threads");
+        LOG_ERR("Failed to start threads.");
         return 1;
     }