+++ src/amd/GpuContext.h
@@ -5,6 +5,7 @@
  * Copyright 2014-2016 Wolf9466    <https://github.com/OhGodAPet>
  * Copyright 2016      Jay D Dee   <jayddee246@gmail.com>
  * Copyright 2017-2018 XMR-Stak    <https://github.com/fireice-uk>, <https://github.com/psychocrypt>
+ * Copyright 2018      SChernykh   <https://github.com/SChernykh>
  * Copyright 2016-2018 XMRig       <https://github.com/xmrig>, <support@xmrig.com>
  *
  *   This program is free software: you can redistribute it and/or modify
@@ -21,15 +22,22 @@
  *   along with this program. If not, see <http://www.gnu.org/licenses/>.
  */
 
-#ifndef __GPUCONTEXT_H__
-#define __GPUCONTEXT_H__
+#ifndef XMRIG_GPUCONTEXT_H
+#define XMRIG_GPUCONTEXT_H
 
 
-#include "3rdparty/CL/cl.h"
+#if defined(__APPLE__)
+#   include <OpenCL/cl.h>
+#else
+#   include "3rdparty/CL/cl.h"
+#endif
 
 
 #include <stdint.h>
 #include <string>
+
+
+#include "common/xmrig.h"
 
 
 struct GpuContext
@@ -41,6 +49,8 @@ struct GpuContext
         stridedIndex(1),
         memChunk(2),
         compMode(1),
+        unrollFactor(8),
+        vendor(xmrig::OCL_VENDOR_UNKNOWN),
         DeviceID(nullptr),
         CommandQueues(nullptr),
         InputBuffer(nullptr),
@@ -54,13 +64,15 @@ struct GpuContext
     {}
 
 
-    inline GpuContext(size_t index, size_t intensity, size_t worksize, int stridedIndex, int memChunk, bool compMode) :
+    inline GpuContext(size_t index, size_t intensity, size_t worksize, int stridedIndex, int memChunk, bool compMode, int unrollFactor) :
         deviceIdx(index),
         rawIntensity(intensity),
         workSize(worksize),
         stridedIndex(stridedIndex),
         memChunk(memChunk),
         compMode(compMode ? 1 : 0),
+        unrollFactor(unrollFactor),
+        vendor(xmrig::OCL_VENDOR_UNKNOWN),
         DeviceID(nullptr),
         CommandQueues(nullptr),
         InputBuffer(nullptr),
@@ -80,6 +92,8 @@ struct GpuContext
     int stridedIndex;
     int memChunk;
     int compMode;
+    int unrollFactor;
+    xmrig::OclVendor vendor;
 
     /*Output vars*/
     cl_device_id DeviceID;
@@ -88,7 +102,7 @@ struct GpuContext
     cl_mem OutputBuffer;
     cl_mem ExtraBuffers[6];
     cl_program Program;
-    cl_kernel Kernels[11];
+    cl_kernel Kernels[12];
     size_t freeMem;
     int computeUnits;
     std::string name;
@@ -97,4 +111,4 @@ struct GpuContext
 };
 
 
-#endif /* __GPUCONTEXT_H__ */
+#endif /* XMRIG_GPUCONTEXT_H */